chore: move metrics from the wrong folder
